+++ entry/src/ohosTest/java/com/github/nikartm/stripedprocessbutton/Util.java
@@ -1,0 +1,31 @@
+package com.github.nikartm.stripedprocessbutton;
+
+import ohos.agp.components.Attr;
+import ohos.agp.components.AttrSet;
+import java.util.Optional;
+
+public class Util {
+    public static AttrSet setUpAttrSet() {
+        return new AttrSet() {
+            @Override
+            public Optional<String> getStyle() {
+                return Optional.empty();
+            }
+
+            @Override
+            public int getLength() {
+                return 0;
+            }
+
+            @Override
+            public Optional<Attr> getAttr(int i) {
+                return Optional.empty();
+            }
+
+            @Override
+            public Optional<Attr> getAttr(String s) {
+                return Optional.empty();
+            }
+        };
+    }
+}
